requirements.txt
langchain>=1.2.10
langchain-openai>=1.1.10
langchain-community>=0.4.1
langchain-text-splitters>=1.0.0
chromadb>=0.5.5
openai>=1.51.0
python-dotenv>=1.0.1
requests>=2.32.3
pypdf>=5.0.1
streamlit>=1.39.0
